refactor(ai-builder): Consolidate all prompts into src/prompts/
- Move chains/prompts/ to src/prompts/chains/parameter-updater/
- Extract chain templates to dedicated .prompt.ts files
- Rename main-agent to legacy-agent for clarity
- Add comprehensive README documenting prompt architecture
