+++ packages/@n8n/ai-workflow-builder.ee/src/chains/conversation-compact.ts
@@ -1,22 +1,9 @@
 import type { BaseChatModel } from '@langchain/core/language_models/chat_models';
 import type { BaseMessage } from '@langchain/core/messages';
 import { AIMessage, HumanMessage } from '@langchain/core/messages';
-import { PromptTemplate } from '@langchain/core/prompts';
 import z from 'zod';
 
-const compactPromptTemplate = PromptTemplate.fromTemplate(
-	`Please summarize the following conversation between a user and an AI assistant building an n8n workflow:
-
-<previous_summary>
-{previousSummary}
-</previous_summary>
-
-<conversation>
-{conversationText}
-</conversation>
-
-Provide a structured summary that captures the key points, decisions made, current state of the workflow, and suggested next steps.`,
-);
+import { compactPromptTemplate } from '@/prompts/chains/compact.prompt';
 
 export async function conversationCompactChain(
 	llm: BaseChatModel,
