+++ packages/@n8n/ai-workflow-builder.ee/src/chains/parameter-updater.ts
@@ -4,10 +4,13 @@ import { ChatPromptTemplate, HumanMessagePromptTemplate } from '@langchain/core/
 import type { Logger } from 'n8n-workflow';
 import { z } from 'zod';
 
+import {
+	instanceUrlPrompt,
+	ParameterUpdatePromptBuilder,
+} from '@/prompts/chains/parameter-updater';
+
 import { LLMServiceError } from '../errors';
 import type { ParameterUpdaterOptions } from '../types/config';
-import { instanceUrlPrompt } from './prompts/instance-url';
-import { ParameterUpdatePromptBuilder } from './prompts/prompt-builder';
 
 export const parametersSchema = z
 	.object({
