+++ packages/@n8n/ai-workflow-builder.ee/src/prompts/chains/parameter-updater/base/common-patterns.ts
