+++ packages/@n8n/ai-workflow-builder.ee/src/prompts/chains/parameter-updater/base/output-format.ts
