+++ packages/@n8n/ai-workflow-builder.ee/src/prompts/chains/parameter-updater/examples/advanced/resource-locator-examples.ts
