+++ packages/@n8n/ai-workflow-builder.ee/src/prompts/chains/parameter-updater/examples/advanced/tool-node-examples.ts
