+++ packages/@n8n/ai-workflow-builder.ee/src/prompts/chains/parameter-updater/examples/basic/if-node-examples.ts
