+++ packages/@n8n/ai-workflow-builder.ee/src/prompts/chains/parameter-updater/examples/basic/set-node-examples.ts
