+++ packages/@n8n/ai-workflow-builder.ee/src/prompts/chains/parameter-updater/examples/basic/switch-node-examples.ts
