+++ packages/@n8n/ai-workflow-builder.ee/src/prompts/chains/parameter-updater/node-types/switch-node.ts
