+++ packages/@n8n/ai-workflow-builder.ee/src/prompts/chains/parameter-updater/parameter-types/text-fields.ts
