+++ packages/@n8n/ai-workflow-builder.ee/src/prompts/chains/parameter-updater/prompt-config.ts
@@ -1,4 +1,4 @@
-import type { NodePromptConfig } from '../../types/config';
+import type { NodePromptConfig } from '@/types/config';
 
 export const DEFAULT_PROMPT_CONFIG: NodePromptConfig = {
 	nodeTypePatterns: {
