+++ packages/@n8n/ai-workflow-builder.ee/src/prompts/chains/parameter-updater/test/prompt-builder.test.ts
